Separated libpng dependency to own file
Separated all code depending on libpng to its own source file, in
preparation for lodepng library integration.
